requirements.txt
Plate==1.0.1
TgCrypto==1.2.5
Pyrogram==2.0.102
python-dotenv==1.0.0
git+https://github.com/NicKoehler/brawlhalla_api@dev
